Improve orig and new
Scan no longer makes a fatal error when encountering a file outside the
orig path prefix

No longer passing empty string to filepath.Abs which returns the current
directory instead of empty.

Providing -in with an argument not under the current directory also
caused this error to be thrown. This was due to filepath.Abs behavior,
so that is also fixed